+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ## [Unreleased]
 ### Added
 - Initial implementation of recursive dependency resolution.
+- The `bender packages` command to access the package dependency graph.
 
 ### Changed
 - Lockfile now contains the dependency names for each package.